Create new node_attributes.go
